src/requirements.txt
bidict==0.22.0
cachetools==5.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-SocketIO==5.1.2
google-auth==2.6.6
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
kubernetes==11.0.0
MarkupSafe==2.1.1
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-engineio==4.3.2
python-socketio==5.6.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
websocket-client==1.3.2
Werkzeug==2.1.2
zipp==3.8.0
